Major redesign of the database
The redesign has 2 parts:
  1. New data model for the database
  2. Use of member functions instead of member variables

The main change in the data model is that now there is a Bug table
with the data shared by all the phases of a bug (eggs, alive bug,
dead bug) and a table for phase. The primary key of the phases is
the identificator of the Bug table.