Avoid infinite loop when doing a permutation of a word
When doing a permutation of a word, if the bytes of the word have
all the same value, all possible results are equal to original
word.

Before 12860ba3 the mutation was added to the database but as
the original and mutated word are the same, the mutation isn't real.

After 12860ba3 a loop was added to make sure a real mutation was
created and in case of a non-permutable word. In the case of a
non-permutable word, the program ends in a infinite loop.

With this commit a mutation of type Total, Partial or Permutation
is tried only once and if the original and mutate words are
the same the mutation isn't added to the database.

This is a revert/fix of 12860ba3a2f8b1d1d94d8b533fdc901554b4d8b5