db: New version of the database (version 6)
Now there is two possible versions of the code of a bug:
- The initial code is the code is the code that a egg has when it's created.
  This code can be the code of a spawn or a copy of the executable code of
  its father. This initial code of a bug isn't never modified.
- The executable code is the code that the bug is executing while it's alive.
  When the bug died, this code is deleted from the database.

As the initial code can't be modified, the bugs spawned will share the same
code. Also, as the executable code is deleted after the death of a bug, the
database size of this version is lower than the previous version.

As the changes to the database were not compatible, the database version was
increased to 6.