Don't try to force to get a different word when a word is mutated
When a permutation over a word is done, if this word has the value
of 0x00000000, the new value will have always the same value as the
old one. So, before this change, a infinite loop was done.
With a total or partial mutation the previous loop was ok, but the
loop was deleted too for consistent behaviour with the permutation.